refactor: abstract functions to custom hook
