+++ frontend/src/components/page/Notas.tsx
@@ -1,68 +1,20 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import NewNoteButton from '../atoms/notes/NewNoteButton'
 import Title from '../atoms/Title'
 import NewNoteModal from '../molecules/notes/NewNoteModal'
 import CancelNoteButton from '../atoms/notes/CancelNoteButton'
 import SaveNoteButton from '../atoms/notes/SaveNoteButton'
-import Note from '@/types/note'
-import patient from '@/types/patient'
+import usePostNotes from '@/hooks/notes/usePostNotes'
 
 function Notas() {
-  const [showModal, setShowModal] = useState(false)
-  const [noteInfo, setNoteInfo] = useState<Note>({
-    general: null,
-    illness: null,
-    recepie: null
-  })
-  const [patients, setPatients] = useState<patient[]>([])
-  const [selectedPatient, setSelectedPatient] = useState<string>("")
-
-  useEffect(() => {
-    fetch("/api/patients/doctorPatients", {
-      credentials: "include"
-    })
-    .then(async res => {
-      const data = await res.json()
-      
-      if (!res.ok) {
-        throw new Error(data.error || "Error desconocido")
-      }
-
-      return data;
-    })
-    .then(data => {
-      const patientsInfo: patient[] = data.map(d => {
-        const name = d.user.name
-        const parentalLastName = d.user.parent_last_name
-        const maternalLastName = d.user.maternal_last_name
-        const userId = d.patient_id
-        
-        return { name, parentalLastName, maternalLastName, userId }
-      })
-
-      setPatients(patientsInfo);
-    })
-    .catch(error => console.error("Error: ", error))
-  }, [])
-
-  function save() {
-    if (noteInfo && selectedPatient) {
-      console.log(JSON.stringify({patientId: selectedPatient, noteInfo}))
-      fetch("/api/notes", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json"
-        },
-        body: JSON.stringify({patientId: selectedPatient, noteInfo}),
-        credentials: "include"
-      })
-      .then((res) => console.log(res))
-      .catch((e) => {
-        console.error(e)
-      })
-      setShowModal(false)
-    }
-  }
+  const {
+    showModal, 
+    patients, 
+    setNoteInfo, 
+    setSelectedPatient, 
+    setShowModal, 
+    save
+  } = usePostNotes();
 
   return (
     <div className='w-1/2'>
